+++ .github/workflows/cicd.yml
@@ -13,29 +13,31 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.7", "3.8", "3.9"]
+        python-version: ["3.7", "3.8", "3.9", "3.10"]
 
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
-      - run: pip install tox tox-gh-actions
+      - run: pip install tox tox-gh-actions poetry
       - run: tox
 
-  coverage:
-    name: Coverage
+  # Disabled until we can introduce pytest and use conventional commands for
+  # gathering coverage with Poetry, pytest, and coverage.py.
+  #coverage:
+  #  name: Coverage
 
-    runs-on: ubuntu-latest
+  #  runs-on: ubuntu-latest
 
-    container:
-      image: python:3.9-buster
+  #  container:
+  #    image: python:3.9-buster
 
-    steps:
-      - uses: actions/checkout@v2
-      - run: pip install coverage codecov
-      - run: coverage run setup.py test
-      - run: coverage report -m
-      - uses: "codecov/codecov-action@v1"
-        with:
-          fail_ci_if_error: true
+  #  steps:
+  #    - uses: actions/checkout@v2
+  #    - run: pip install coverage codecov poetry
+  #    - run: poetry run coverage run -m pytest
+  #    - run: poetry run coverage report -m
+  #    - uses: "codecov/codecov-action@v1"
+  #      with:
+  #        fail_ci_if_error: true
