+++ pyproject.toml
@@ -29,10 +29,13 @@ packages = [
     { include = "django_scim", from = "src" },
 ]
 
+[tool.poetry.scripts]
+test = 'scripts:run_tests'
+
 [tool.poetry.dependencies]
 python = ">=3.6"
 scim2-filter-parser = ">=0.3.5"
-Django = ">=2.0"
+django = ">=3.2"
 python-dateutil = ">=2.7.3"
 
 [tool.poetry.dev-dependencies]
@@ -45,46 +48,7 @@ flake8-isort = "^3.0.0"
 [tool.black]
 line-length = 100
 skip-string-normalization = true
-[tool.tox]
-legacy_tox_ini = """
-[tox]
-envlist =
-    py{37,38,39,310}-django{22,31,32}
-    py{38,39,310}-django{40}
-    flake8
-
-[gh-actions]
-python =
-    3.7: py37
-    3.8: py38
-    3.9: py39
-    3.10: py310, flake8
-
-[testenv]
-setenv =
-    PYTHONDONTWRITEBYTECODE=1
-    DJANGO_SETTINGS_MODULE=test_settings
-basepython =
-    py37: python3.7
-    py38: python3.8
-    py39: python3.9
-    py310: python3.10
-deps =
-    django22: Django>=2.2,<3
-    django31: Django>=3.1,<3.2
-    django32: Django>=3.2,<4
-    django40: Django>=4.0
-commands = python setup.py test
-
-[testenv:flake8]
-basepython =
-    python3.10
-deps =
-    flake8
-commands =
-    flake8 src
-"""
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
