+++ tox.ini
@@ -1,0 +1,39 @@
+[tox]
+isolated_build = True
+envlist =
+    py{37,38,39,310}-django{32}
+    py{38,39,310}-django{40}
+    flake8
+
+[gh-actions]
+python =
+    3.7: py37
+    3.8: py38
+    3.9: py39
+    3.10: py310, flake8
+
+[testenv]
+setenv =
+    PYTHONDONTWRITEBYTECODE=1
+    DJANGO_SETTINGS_MODULE=tests.settings
+basepython =
+    py37: python3.7
+    py38: python3.8
+    py39: python3.9
+    py310: python3.10
+    .package: python3
+deps =
+    django32: Django>=3.2,<4
+    django40: Django>=4.0
+whitelist_externals = poetry
+commands =
+    poetry install -v
+    poetry run test
+
+[testenv:flake8]
+basepython =
+    python3.10
+deps =
+    flake8
+commands =
+    flake8 src
